Add option to show a google map for a location
Add a 'show map' option to the location part which will render out a
google map as another shape to the content item.

Update the LocationPartDriver to render out the new Parts.Location.Map
shape as well as the default location shape when displaying.

Add a new shape for rendering a map based on the location part coordinate
values.

Add a settings page that allows the user to enter their Google Maps API
key for use with the Javascript API. Add a handler which attaches the
settings to the site context item and creates the settings area in the
admin menu.

Add javascript that will handle the callback from google maps api for each
content item on the page with a map and show the location specified with a
pin displayed.

Add basic style to ensure the map is visible in the default theme.