Add a MapLink shape
Add a new MapLink shape which by default displays in summary meta.
Includes a 'show on map' link to the location specified by the
LocationPart which will open in a new window.

Add GoogleMapUrlService which will produce a google maps search result
based on the coordinates if they are available, otherwise based on the
full address.

Add a Show map link option to the LocationPart editor.