Add spacial search settings
Add a new settings area for the spacial search feature which allows admin
users to configure how spacial search is applied. Settings include
enable/disable spacial search filters, spacial search radius and spacial
search filter weighting. These settings allow front end tweaking of the
behaviour of spacial search.

Add a new dependency ILocationSearchService that exposes these settings as
a property, allowing the search implementor to hook in to the values
configured in the UI if they wish.

Update the version number to 0.4.

Update search builder extensions to allow the spacial search filters to
not be mandatory and to allow configuration of the radius and weighting of
them.

Add a new method to search builder to allow passing of multiple
LocationResult values instead of just 1 so that you can pass all results
of geocoding to the search.